Expanded the README to be a basic manual.
Now plot_results.R generates a pdf.

Replaced final.dat with num_final.dat and zbar_final.dat so that output can
   be directly used as input.

Fixed bug where start_t wasn't doing anything.
Made start_t, stop_t, and record_interval be integers.
Better error checking in record-related input parameters.

Removed old/new swapping.  This was possible because most of the old/new
   notation is now contained in competition_happens() and dispersal_happens().

In dispersal_happens, don't bother if delta = 0.

+++ src/input.h
@@ -27,9 +27,9 @@ typedef struct
 	const char *initial_zbar;       /* ...for n and zbar                */
 
 	/* record-keeping */
-	double start_t;
-	double stop_t;
-	double record_interval;
+	int start_t;
+	int stop_t;
+	int record_interval;
 
 } Params;
